pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leenaai-firestoredb"
version = "1.0.0"
description = "Firestore and ChromaDB package for LeenaAI real estate chatbot"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "LeenaAI Team", email = "info@leenaai.com"}
]
maintainers = [
    {name = "LeenaAI Team", email = "info@leenaai.com"}
]
keywords = ["firestore", "chroma", "vector", "real-estate", "chatbot", "database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "google-cloud-firestore>=2.11.0",
    "google-cloud-storage>=2.10.0",
    "google-auth>=2.17.0",
    "google-api-python-client>=2.95.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.0.0",
    "pandas>=1.5.0",
    "numpy>=1.24.0",
    "aiohttp>=3.8.0",
    "asyncio-mqtt>=0.11.0",
    "python-dotenv>=1.0.0",
    "pytz>=2023.3",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/leenaai/leenaai-firestoredb"
Documentation = "https://github.com/leenaai/leenaai-firestoredb/blob/main/README.md"
Repository = "https://github.com/leenaai/leenaai-firestoredb"
"Bug Tracker" = "https://github.com/leenaai/leenaai-firestoredb/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
] 